requirements.txt
# Dépendances Python pour le backend
# Compatible avec Python 3.8+
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Session==0.5.0
pymongo==4.6.0
bcrypt==4.1.1
python-dotenv==1.0.0
email-validator==2.1.0
Werkzeug==3.0.1
Jinja2==3.1.2
reportlab==4.0.7
Pillow>=10.4.0
python-dateutil==2.8.2
pytz==2024.1
gunicorn==21.2.0
requests==2.31.0

